deprecate fanoutv1
